8294549: configure script should detect unsupported path
The OpenJDK build system does not support building when the source code
resides on a path that contains a space. This requirement is documented in the
build instructions but not enforced by the configure script.

This change adds an explicit checks to the wrapper `configure` script that fail
the build if the source code to be built is located in a directory who's path
contains a space character or the build path cannot be determined.